feat: 아이템 검색할 때 keyword를 입력하지 않아도 조회할 수 있도록 기능 추가

+++ lime-api/src/main/java/com/programmers/lime/domains/item/application/dto/ItemGetByCursorServiceResponse.java
@@ -5,7 +5,7 @@ import com.programmers.lime.domains.item.model.ItemCursorSummary;
 
 public record ItemGetByCursorServiceResponse(
 
-	int itemTotalCount,
+	Long itemTotalCount,
 	CursorSummary<ItemCursorSummary> cursorSummary
 ) {
 }
